Experimental branch to mirror Wyriki structure
I am attempting to synchronize my understanding of how the application
will need to flex with how Jim Weirich was teasing apart the different
layers of concern.

His repository layer, renamed to the HydramataServices in this commit,
serves the purpose of the entry point into the application. It is this
class that ultimately does quite a bit of the work.

However, I have been pursuing the Form concept and that also provides
a place to do much of the work. So for now, this branch shall be an
experiment. Perhaps a premature extraction, but one that I believe
will hold value and may be necessitated by the architecture.